pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcptemplate"
version = "0.1.0"
description = "Whatsapp Agent Template"
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "langchain_community>=0.3.17,<0.4",
    "langchain_openai>=0.3.10,<0.4",
    "langchain-google-genai>=2.1.4,<3",
    "langgraph>=0.3.20,<0.4",
    "langgraph-sdk>=0.1.61, <1",
    "langchain-mcp-adapters>=0.0.5,<1",
    "langgraph_supervisor>=0.0.16,<0.1",
    "jinja2>=3.1.6,<4",
    "twilio>=9.5.1, <10",
    "fastapi>=0.115.12",
    "python-multipart>=0.0.20, <0.1.0",
    "uvicorn[standard]>=0.29.0,<0.30",
    "requests>=2.31.0,<3",
    "python-dotenv>=1.0.0,<2",
    "llama-index>=0.12.35,<1",
    "weaviate-client>=4.14.3, <5",
    "langchain-arcade"
]

[[project.authors]]
name = "lgesuellip"
email = "lautaro@pampa.ai"

[tool.poetry]
packages = [
    {include = "agents", from = "src"},
    {include = "langgraph_whatsapp", from = "src"}
]

[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/langgraph_whatsapp"]

[project.scripts]
react-agent = "agents.base.react:main"
